+++ src/main/java/org/samsara/pistachio/interceptor/AdminLoginInterceptor.java
@@ -1,0 +1,35 @@
+package org.samsara.pistachio.interceptor;
+
+import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+/**
+ * Samsara Pistachio
+ * Admin login Interceptor
+ * @author sczyh30
+ * @since 0.1.42
+ */
+public class AdminLoginInterceptor extends HandlerInterceptorAdapter {
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
+        if(request.getSession().getAttribute("auth_key") != null) {
+            response.sendRedirect("../adm1n666/index.freed0m");
+            return false;
+        }
+        return super.preHandle(request, response, handler);
+    }
+
+    @Override
+    public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler, ModelAndView modelAndView) throws Exception {
+        super.postHandle(request, response, handler, modelAndView);
+    }
+
+    @Override
+    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) throws Exception {
+        super.afterCompletion(request, response, handler, ex);
+    }
+}