+++ src/main/java/org/samsara/pistachio/mapper/BookDetailMapper.java
@@ -9,15 +9,31 @@ import org.samsara.pistachio.entity.BookDetail;
  */
 public interface BookDetailMapper {
 
+    /**
+     * WARNING: THIS PROCESS SHOULD BE CAREFULLY!<br>
+     * Remove a book detail entry from the database
+     * @param ISBN ISBN of the book
+     * @return the status
+     */
     boolean remove(String ISBN);
 
-    int insert(BookDetail record);
+    /**
+     * Insert a book detail entity to the database
+     * @param record book detail entity
+     * @return the status
+     */
+    boolean insert(BookDetail record);
 
-    int insertSelective(BookDetail record);
+    boolean insertSelective(BookDetail record);
 
-    BookDetail get(String ISBN);
+    /**
+     * Get the book detail entity by ISBN
+     * @param ISBN ISBN
+     * @return the entity
+     */
+    BookDetail getDetail(String ISBN);
 
-    int updateByPrimaryKeySelective(BookDetail record);
+    boolean updateByPrimaryKeySelective(BookDetail record);
 
-    int update(BookDetail record);
+    boolean update(BookDetail record);
 }