+++ src/main/java/org/samsara/pistachio/service/BookSearchService.java
@@ -18,11 +18,49 @@ public class BookSearchService {
     @Resource
     private BookInfoMapper infoMapper;
 
+    public BookInfo getByName(String name) {
+        return infoMapper.getByName(name);
+    }
+
+    /**
+     * Get all books by author name
+     * @param author distinct name of author
+     * @return the query result
+     */
     public List<BookInfo> getAllBooksByAuthor(String author) {
         return infoMapper.getAllByAuthor(author);
     }
-
-    public List<String> getAllNameByAuthor(String author) {
+    /**
+     * Get all names of book by author name
+     * @param author distinct name of author
+     * @return the query result
+     */
+    public List<String> getAllNamesByAuthor(String author) {
         return infoMapper.getAllNameByAuthor(author);
+    }
+
+    // TODO: NEED HIGHER PERFORMANCE!!!!
+    /**
+     * Get all books by author name (vague query)<br>
+     * WARNING: this process could consume high because it can
+     * make use of the index of the database.<br>
+     * Next version, we will use Solr to achieve this
+     * @param vague vague name of author
+     * @return the query result
+     */
+    public List<BookInfo> getAllBooksByVgAuthor(String vague) {
+        return infoMapper.getAllByVgAuthor(vague);
+    }
+
+    /**
+     * Get all books by book name (vague query)<br>
+     * WARNING: this process could consume high because it can
+     * make use of the index of the database.<br>
+     * Next version, we will use Solr to achieve this
+     * @param vague vague name of book
+     * @return the query result
+     */
+    public List<BookInfo> getAllBooksByVgName(String vague) {
+        return infoMapper.getAllByVgName(vague);
     }
 }