+++ src/main/java/org/samsara/pistachio/service/BookService.java
@@ -1,11 +1,22 @@
 package org.samsara.pistachio.service;
 
 import org.samsara.pistachio.entity.BookInfo;
+import org.samsara.pistachio.entity.BookStatus;
+import org.samsara.pistachio.entity.ProcessStatus;
 import org.samsara.pistachio.mapper.BookInfoMapper;
+import org.samsara.pistachio.mapper.BookStatusMapper;
+import org.samsara.pistachio.mapper.ProcessStatusMapper;
+import org.samsara.pistachio.security.TokenGenerator;
+import org.samsara.pistachio.util.DateUtil;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
+import java.sql.Timestamp;
+import java.util.Date;
+import java.util.List;
+
+import static org.samsara.pistachio.Constant.*;
 
 /**
  * Samsara Pistachio Service
@@ -18,17 +29,51 @@ public class BookService {
     @Resource
     private BookInfoMapper infoMapper;
 
+    @Resource
+    private BookStatusMapper statusMapper;
+
+    @Resource
+    private ProcessStatusMapper processStatusMapper;
+
+    public ProcessStatus retAndRecord(int code, int id) {
+        ProcessStatus status = wrap(code, id);
+        processStatusMapper.insert(status);
+        return status;
+    }
+
+    public ProcessStatus wrap(int code, int id) {
+        String psid = TokenGenerator.generate(667788, code); // 667788 is a magic number
+        Timestamp timestamp = DateUtil.getStamp();
+        System.out.println(timestamp);
+        switch (code) {
+            case BOOK_INFO_REMOVE_SUCCESS:
+                return new ProcessStatus(code, psid, "book_info_remove_success", API_PROCESS_LEVEL_ADMIN, id, timestamp);
+            case BOOK_INFO_ADD_SUCCESS:
+                return new ProcessStatus(code, psid, "book_info_add_success", API_PROCESS_LEVEL_ADMIN, id, timestamp);
+            case BOOK_INFO_UPDATE_SUCCESS:
+                return new ProcessStatus(code, psid, "book_info_update_success", API_PROCESS_LEVEL_ADMIN, id, timestamp);
+            case BOOK_INFO_PROCESS_FAILURE:
+                return new ProcessStatus(code, psid, "book_info_process_error", API_PROCESS_LEVEL_ADMIN, id, timestamp);
+            default:
+                return new ProcessStatus(4606, psid, "book_process_unknown", API_PROCESS_LEVEL_NO, id, timestamp);
+        }
+    }
+
+    public BookStatus getStatus(String ISBN) {
+        return statusMapper.getStatus(ISBN);
+    }
+
     /**
      * Add a book to the database
      * @param bookInfo the BookInfo entity
      * @return true if insert process is successful; else false
      */
-    public boolean addBook(BookInfo bookInfo) {
-        return infoMapper.insert(bookInfo);
+    public int addBook(BookInfo bookInfo) {
+        return infoMapper.insert(bookInfo) ? BOOK_INFO_ADD_SUCCESS : BOOK_INFO_PROCESS_FAILURE;
     }
 
     /**
-     * Get the basic book info by ISBN
+     * Get the book info by ISBN
      * @param ISBN ISBN of the book
      * @return the BookInfo entity
      */
@@ -38,7 +83,16 @@ public class BookService {
     }
 
     /**
-     * Get the basic book info by book name
+     * Get latest books
+     * @return list of the books
+     */
+    @Transactional(readOnly = true)
+    public List<BookInfo> getLatestBook() {
+        return infoMapper.getLatest();
+    }
+
+    /**
+     * Get the book info by book name
      * @param name name of the book
      * @return the BookInfo entity
      */
@@ -48,12 +102,22 @@ public class BookService {
     }
 
     /**
+     * Get all books by the author
+     * @param author the author name
+     * @return list of the books
+     */
+    @Transactional(readOnly = true)
+    public List<BookInfo> getAllBooksByAuthor(String author) {
+        return infoMapper.getAllByAuthor(author);
+    }
+
+    /**
      * Update the book info by ISBN
      * @param bookInfo the BookInfo entity
      * @return true if update process is successful; else false
      */
-    public boolean updateBook(BookInfo bookInfo) {
-        return infoMapper.update(bookInfo);
+    public int updateBook(BookInfo bookInfo) {
+        return infoMapper.update(bookInfo) ? BOOK_INFO_UPDATE_SUCCESS : BOOK_INFO_PROCESS_FAILURE;
     }
 
     /**
@@ -61,10 +125,10 @@ public class BookService {
      * Because the database correlation map is <strong>cascade</strong>, the book detail entity,
      * book status entity will automatically be removed.
      * @param ISBN ISBN of the book
-     * @return true if remove process is successful; else false
+     * @return the status code
      */
-    public boolean removeBook(String ISBN) {
-        return infoMapper.remove(ISBN);
+    public int removeBook(String ISBN) {
+        return infoMapper.remove(ISBN) ? BOOK_INFO_REMOVE_SUCCESS : BOOK_INFO_PROCESS_FAILURE;
     }