+++ src/main/java/org/samsara/pistachio/service/UserService.java
@@ -23,4 +23,5 @@ public class UserService {
     public User getUser(String username) {
         return userMapper.getByName(username);
     }
+
 }